+++ pybryt/annotations/__init__.py
@@ -2,5 +2,6 @@
 
 from . import invariants
 from .annotation import Annotation, AnnotationResult
+from .complexity import *
 from .relation import *
 from .value import *